Adicionando collection cache na lista de preços

+++ app/jobs/book_price_search_job.rb
@@ -4,6 +4,8 @@ class BookPriceSearchJob < ApplicationJob
   def perform(book_id, user_id)
     book = Book.find(book_id)
     search_prices = ThirdPartySearch::ZoomSearchService.call(book.title)
+    return if search_prices.blank?
+
     book.search_prices.destroy_all
     SearchPrice.insert_all(search_prices.map { it.merge(book_id: book.id) })
     SearchPricesMailer.list_prices(user_id, book_id).deliver_later
